fix: bump qstash and review
