+++ src/context/api/anthropic.ts
@@ -30,7 +30,7 @@ type ChatCompletion = {
 };
 
 export class AnthropicAPI extends BaseWorkflowApi {
-  public async call<TBody = CreateChatCompletion, TResponse = ChatCompletion>(
+  public async call<TResult = ChatCompletion, TBody = CreateChatCompletion>(
     stepName: string,
     settings: ApiCallSettings<
       TBody,
@@ -39,10 +39,10 @@ export class AnthropicAPI extends BaseWorkflowApi {
         operation: "messages.create";
       }
     >
-  ): Promise<CallResponse<TResponse>> {
+  ): Promise<CallResponse<TResult>> {
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
     const { token, operation, ...parameters } = settings;
-    return await this.callApi<TResponse>(stepName, {
+    return await this.callApi<TResult, TBody>(stepName, {
       api: {
         name: "llm",
         provider: anthropic({ token }),
